+++ web/modules/custom/os2forms_queued_email/src/Plugin/WebformElement/Os2FormsWebformAudioFile.php
@@ -2,45 +2,15 @@
 
 namespace Drupal\os2forms_queued_email\Plugin\WebformElement;
 
-use Drupal\os2forms_queued_email\Plugin\AdvancedQueue\JobType\QueuedEmail;
 use Drupal\webform\Plugin\WebformElement\WebformAudioFile;
-use Drupal\webform\WebformSubmissionInterface;
 
 /**
- * Overrides 'webform_audio_file' element.
+ * OS2Forms webform audio file.
  *
- * @WebformElement(
- *   id = "os2forms_webform_audio_file",
- *   label = @Translation("Audio file"),
- *   description = @Translation("Provides a form element for uploading and saving an audio file."),
- *   category = @Translation("File upload elements"),
- *   states_wrapper = TRUE,
- *   dependencies = {
- *     "file",
- *   }
- * )
+ * @see Os2FormsQueuedEmailFileTrait
  */
 class Os2FormsWebformAudioFile extends WebformAudioFile {
 
-  /**
-   * {@inheritdoc}
-   *
-   * OS2Forms werbform audio file element.
-   *
-   * Copied from WebformManagedFileBase::getEmailAttachments.
-   * Only change is that filecontent is not passed on.
-   *
-   * @see parent::getEmailAttachments()
-   */
-  public function getEmailAttachments(array $element, WebformSubmissionInterface $webform_submission, array $options = []) {
-    $attachments = parent::getEmailAttachments($element, $webform_submission, $options);
-
-    foreach ($attachments as &$attachment) {
-      $attachment[QueuedEmail::FILECONTENT] = '';
-      $attachment[QueuedEmail::OS2FORMS_QUEUED_EMAIL_IS_STATIC_FILE] = TRUE;
-    }
-
-    return $attachments;
-  }
+  use Os2FormsQueuedEmailFileTrait;
 
 }
